Revert "update"
This reverts commit 1fa2db8a8051796f68f5dc747565b93e4d671007.

+++ datafusion/core/src/test_util/parquet.rs
@@ -102,7 +102,7 @@ impl TestParquetFile {
 
         println!("Generated test dataset with {num_rows} rows");
 
-        let size = std::fs::metadata(&path)?.len();
+        let size = std::fs::metadata(&path)?.len() as usize;
 
         let mut canonical_path = path.canonicalize()?;
